pyproject.toml
[tool.poetry]
name = "tarka"
version = "0.1.0"
description = "Incident investigation agent that converts Prometheus/Alertmanager alerts into actionable triage reports"
authors = ["Dinesh Auti <auti.dinesh3@gmail.com>"]
readme = "README.md"
packages = [{include = "agent"}]

[tool.poetry.dependencies]
python = "^3.12"
kubernetes = ">=28.1.0"
requests = ">=2.31.0"
python-dateutil = ">=2.8.0"
boto3 = ">=1.28.0"
fastapi = ">=0.110.0"
uvicorn = ">=0.29.0"
PyJWT = {extras = ["crypto"], version = ">=2.8.0"}
itsdangerous = ">=2.2.0"
authlib = "^1.3.0"
bcrypt = "^4.0.0"
pydantic = "2.12.5"
psycopg = {extras = ["binary"], version = "3.2.10"}
nats-py = "2.9.0"
langchain = "0.3.27"
langchain-core = "0.3.83"
langgraph = "0.6.11"
langchain-google-vertexai = {version = "2.1.2", optional = true}
langchain-anthropic = {version = "^0.3.0", optional = true}

[tool.poetry.extras]
vertex = ["langchain-google-vertexai"]
anthropic = ["langchain-anthropic"]
all-providers = ["langchain-google-vertexai", "langchain-anthropic"]

[tool.poetry.group.dev.dependencies]
pytest = "8.3.4"
pytest-cov = "5.0.0"
coverage = "7.6.10"
black = "^26.1.0"
pre-commit = "^4.5.1"
isort = "^7.0.0"
flake8 = "^7.3.0"
mypy = "^1.19.1"
autoflake = "^2.3.1"
pytest-asyncio = "^1.3.0"
pytest-html = "^4.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-q -m 'not e2e'"
markers = [
  "integration: requires external services (e.g., NATS JetStream) and is run in CI integration job",
  "e2e: end-to-end tests requiring full stack (webhook, worker, UI) - run in CI e2e job or manually",
]

[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
include_trailing_comma = true

[tool.mypy]
install_types = true
non_interactive = true
strict = true
plugins = "pydantic.mypy"

[tool.flake8]
max-line-length = 120
exclude = "ui/node_modules"

[tool.black]
line-length = 120
